pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "docmill"
version = "0.1.0"
description = "A unified inference runtime for OCR & VLM document understanding."
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "DocMill Contributors"},
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["docmill*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
